+++ src/__tests__/jest-dom.finished.js
@@ -1,0 +1,11 @@
+import 'jest-dom/extend-expect'
+import React from 'react'
+import ReactDOM from 'react-dom'
+import {FavoriteNumber} from '../favorite-number'
+
+test('renders a number input with a label "Favorite Number"', () => {
+  const div = document.createElement('div')
+  ReactDOM.render(<FavoriteNumber />, div)
+  expect(div.querySelector('input')).toHaveAttribute('type', 'number')
+  expect(div.querySelector('label')).toHaveTextContent('Favorite Number')
+})
