+++ src/__tests__/react-dom.js
@@ -1,10 +1,11 @@
-import React from 'react'
-import ReactDOM from 'react-dom'
-import {FavoriteNumber} from '../favorite-number'
+// 🐨 you're going to need to use React to create react elements, so import react
+// 🐨 we're going to render the FavoriteNumber component with ReactDOM so you'll need to import react-dom
+// 🐨 Here's your component:
+// import {FavoriteNumber} from '../favorite-number'
 
-test('renders a number input with a label "Favorite Number"', () => {
-  const div = document.createElement('div')
-  ReactDOM.render(<FavoriteNumber />, div)
-  expect(div.querySelector('input').type).toBe('number')
-  expect(div.querySelector('label').textContent).toBe('Favorite Number')
+test.skip('renders a number input with a label "Favorite Number"', () => {
+  // 🐨 create a div (💯 document.createElement)
+  // 🐨 render the FavoriteNumber component to that div with ReactDOM.render
+  // 🐨 assert the input type attribute is a number
+  //:🐨 assert the label's text content is "Favorite Number"
 })
