+++ src/__tests__/react-testing-library.finished.js
@@ -1,0 +1,11 @@
+import 'jest-dom/extend-expect'
+import 'react-testing-library/cleanup-after-each'
+import React from 'react'
+import {render} from 'react-testing-library'
+import {FavoriteNumber} from '../favorite-number'
+
+test('renders a number input with a label "Favorite Number"', () => {
+  const {getByLabelText} = render(<FavoriteNumber />)
+  const input = getByLabelText(/favorite number/i)
+  expect(input).toHaveAttribute('type', 'number')
+})
